Numbers, and using them to estimate other numbers:
- Estimate the number of active sites that don't send an ID
- Estimate the frequency of update checks

According to the results of this code, in order to accurately count
at least 99% of sites, an update check should cause the site to show up
as "active" for around 7 to 10 days.

Lots of assumptions in here that things are normally distributed, but
this should be good enough to get a ballpark value.